review fixes: deduplicate propagate_struct_nulls, fix docs and Python
- Extract propagate_struct_nulls to crate-level shared function,
  removing duplication between writer.rs and bucket_writer.rs.
- Fix complex-types.html: update __null__ naming from __null__(<id>)
  to name.__null__ format, fix expanded column name examples.
- Fix Python _build_projected_schema: implement exact-match-first
  for top-level column names containing '.', consistent with Rust.
