+++ core/src/bucket_writer.rs
@@ -319,27 +319,7 @@ impl BucketWriter {
     }
 
     fn propagate_struct_nulls(struct_arr: &arrow_array::StructArray, child: &ArrayRef) -> ArrayRef {
-        use arrow_buffer::{BooleanBuffer, Buffer, NullBuffer};
-        let num_rows = struct_arr.len();
-        let mut null_bm = vec![0xFFu8; num_rows.div_ceil(8)];
-
-        for i in 0..num_rows {
-            let valid = !child.is_null(i) && !struct_arr.is_null(i);
-            if !valid {
-                null_bm[i / 8] &= !(1 << (i % 8));
-            }
-        }
-
-        let new_null_buf =
-            NullBuffer::new(BooleanBuffer::new(Buffer::from_vec(null_bm), 0, num_rows));
-        arrow_array::make_array(
-            child
-                .to_data()
-                .into_builder()
-                .null_bit_buffer(Some(new_null_buf.into_inner().into_inner()))
-                .build()
-                .unwrap(),
-        )
+        crate::propagate_struct_nulls(struct_arr, child)
     }
 
     fn append_array_column(
