+++ core/src/writer.rs
@@ -285,32 +285,8 @@ impl<S: OutputFile> MosaicWriter<S> {
         }
     }
 
-    /// Propagate STRUCT nulls to a child array: where struct is null, child becomes null.
-    pub(crate) fn propagate_struct_nulls(
-        struct_arr: &arrow_array::StructArray,
-        child: &ArrayRef,
-    ) -> ArrayRef {
-        use arrow_buffer::{BooleanBuffer, Buffer, NullBuffer};
-        let num_rows = struct_arr.len();
-        let mut null_bm = vec![0xFFu8; num_rows.div_ceil(8)];
-
-        for i in 0..num_rows {
-            let valid = !child.is_null(i) && !struct_arr.is_null(i);
-            if !valid {
-                null_bm[i / 8] &= !(1 << (i % 8));
-            }
-        }
-
-        let new_null_buf =
-            NullBuffer::new(BooleanBuffer::new(Buffer::from_vec(null_bm), 0, num_rows));
-        arrow_array::make_array(
-            child
-                .to_data()
-                .into_builder()
-                .null_bit_buffer(Some(new_null_buf.into_inner().into_inner()))
-                .build()
-                .unwrap(),
-        )
+    fn propagate_struct_nulls(struct_arr: &arrow_array::StructArray, child: &ArrayRef) -> ArrayRef {
+        crate::propagate_struct_nulls(struct_arr, child)
     }
 
     fn build_batch_col_map(schema: &Schema, mosaic: &MosaicSchema) -> Vec<usize> {
